+++ examples/order_process_app.rs
@@ -2,13 +2,12 @@
 extern crate serde_derive;
 
 use atomic_counter::{AtomicCounter, RelaxedCounter};
-use futures::stream::Stream;
-use futures::Future;
+use futures::compat::Stream01CompatExt;
+use futures::prelude::*;
+
 use std::sync::Arc;
-use std::thread::sleep;
 use std::time::Duration;
 use structopt::StructOpt;
-use tokio::timer::Interval;
 use zeebest::{Client, JobResult, PanicOption, PublishMessage, WorkflowInstance, WorkflowVersion};
 
 #[derive(StructOpt, Debug)]
@@ -25,7 +24,10 @@ enum Opt {
         name = "place-order",
         about = "Place a new order. This starts a workflow instance."
     )]
-    PlaceOrder,
+    PlaceOrder {
+        #[structopt(short = "c", long = "count")]
+        count: i32,
+    },
     #[structopt(
         name = "notify-payment-received",
         about = "Indicate that the order was processed and there is now a cost for the order."
@@ -56,31 +58,32 @@ struct Payment {
 }
 
 fn main() {
-    let mut client = Client::new("127.0.0.1", 26500).expect("Could not connect to broker.");
+    let client = Client::new("127.0.0.1", 26500).expect("Could not connect to broker.");
 
     let opt = Opt::from_args();
+
     match opt {
         Opt::DeployWorkflow => {
-            client
-                .deploy_bpmn_workflow(
-                    "order-process",
-                    include_bytes!("../examples/order-process.bpmn").to_vec(),
-                )
-                .wait()
-                .unwrap();
+            futures::executor::block_on(client.deploy_bpmn_workflow(
+                "order-process",
+                include_bytes!("../examples/order-process.bpmn").to_vec(),
+            ))
+            .unwrap();
         }
-        Opt::PlaceOrder => {
-            client
-                .create_workflow_instance(WorkflowInstance::workflow_instance_with_bpmn_process(
-                    "order-process",
-                    WorkflowVersion::Latest,
+        Opt::PlaceOrder { count } => {
+            for _ in 0..count {
+                futures::executor::block_on(client.create_workflow_instance(
+                    WorkflowInstance::workflow_instance_with_bpmn_process(
+                        "order-process",
+                        WorkflowVersion::Latest,
+                    ),
                 ))
-                .wait()
                 .unwrap();
+            }
         }
         Opt::NotifyPaymentReceived { order_id, cost } => {
-            client
-                .publish_message(
+            futures::executor::block_on(
+                client.publish_message(
                     PublishMessage::new(
                         "payment-received",
                         order_id.to_string().as_str(),
@@ -89,77 +92,68 @@ fn main() {
                     )
                     .variables(&Payment { order_value: cost })
                     .unwrap(),
-                )
-                .wait()
-                .unwrap();
+                ),
+            )
+            .unwrap();
         }
         Opt::ProcessJobs => {
             let order_id_counter = Arc::new(RelaxedCounter::new(0));
 
-            // a payment worker that passes along a new order_id
-            let mut initiate_payment_worker = client.worker(
-                "rusty-worker",
-                "initiate-payment",
-                Duration::from_secs(3),
-                4,
-                PanicOption::FailJobOnPanic,
-                move |_| {
-                    sleep(Duration::from_secs(5));
-                    Ok({
-                        // increment the order id counter
-                        // this would normally be a key in a database or something
-                        let order_id = order_id_counter.inc();
-                        let variables = serde_json::to_string(&Order {
-                            order_id: order_id as i32,
-                        })
-                        .unwrap();
-                        JobResult::Complete {
-                            variables: Some(variables),
-                        }
-                    })
-                },
-            );
-
-            let mut ship_without_insurance_worker = client.worker(
-                "rusty-worker",
-                "ship-without-insurance",
-                Duration::from_secs(3),
-                4,
-                PanicOption::FailJobOnPanic,
-                |_| Ok(JobResult::Complete { variables: None }),
-            );
-
-            let mut ship_with_insurance_worker = client.worker(
-                "rusty-worker",
-                "ship-with-insurance",
-                Duration::from_secs(3),
-                4,
-                PanicOption::FailJobOnPanic,
-                |_| Ok(JobResult::Complete { variables: None }),
-            );
-
-            let f = Interval::new_interval(Duration::from_secs(1))
-                .map_err(|_| panic!("The interval panicked."))
-                .and_then(move |_| {
-                    let s1 = initiate_payment_worker.activate_and_process_jobs();
-                    let s2 = ship_without_insurance_worker.activate_and_process_jobs();
-                    let s3 = ship_with_insurance_worker.activate_and_process_jobs();
-                    let sa = s1.select(s2);
-                    let sb = sa.select(s3);
-                    sb.map(|jr| {
-                        println!("job result: {:?}", jr);
-                    })
-                    .map_err(|e| {
-                        eprintln!("job errored: {:?}", e);
-                    })
-                    .collect()
-                    .map(|_| ())
+            let initial_payment_handler = move |_| {
+                let order_id_counter = order_id_counter.clone();
+                let order_id = order_id_counter.inc();
+                let variables = serde_json::to_string(&Order {
+                    order_id: order_id as i32,
                 })
-                .collect()
-                .map(|_| println!("Done."))
-                .map_err(|_| ());
+                .unwrap();
+                let job_result = JobResult::Complete {
+                    variables: Some(variables),
+                };
+                futures::future::ready(job_result).boxed()
+            };
 
-            tokio::run(f);
+            let initiate_payment_job = zeebest::JobWorker::new(
+                "rusty-worker".to_string(),
+                "initiate-payment".to_string(),
+                Duration::from_secs(3).as_secs() as _,
+                1,
+                PanicOption::FailJobOnPanic,
+                client.clone(),
+                initial_payment_handler,
+            );
+
+            let ship_without_insurance_job = zeebest::JobWorker::new(
+                "rusty-worker".to_string(),
+                "ship-without-insurance".to_string(),
+                Duration::from_secs(3).as_secs() as _,
+                1,
+                PanicOption::FailJobOnPanic,
+                client.clone(),
+                |_| futures::future::ready(JobResult::Complete { variables: None }).boxed(),
+            );
+
+            let ship_with_insurance_job = zeebest::JobWorker::new(
+                "rusty-worker".to_string(),
+                "ship-with-insurance".to_string(),
+                Duration::from_secs(3).as_secs() as _,
+                1,
+                PanicOption::FailJobOnPanic,
+                client.clone(),
+                |_| futures::future::ready(JobResult::Complete { variables: None }).boxed(),
+            );
+
+            let interval = tokio::timer::Interval::new_interval(Duration::from_secs(4));
+
+            let f = interval.compat().for_each(|_| {
+                let s1 = initiate_payment_job.clone().activate_and_process_jobs();
+                let s2 = ship_with_insurance_job.clone().activate_and_process_jobs();
+                let s3 = ship_without_insurance_job
+                    .clone()
+                    .activate_and_process_jobs();
+                futures::future::join3(s1, s2, s3).then(|_| futures::future::ready(()))
+            });
+
+            futures::executor::block_on(f);
         }
     }
 }