+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- the client and worker now use futures 0.3
+- the worker api has been changed to return a unit future
+- switch to zeebe 0.20 and change to their community license
 
 ## [0.18.1] - 2019-06-18
 ### Added